% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.ece.R
\name{calc.ece}
\alias{calc.ece}
\title{Calculate empirical cross entropy (ECE)}
\usage{
calc.ece(LR.ss, LR.ds, prior = seq(from = 0.01, to = 0.99, length = 99))
}
\arguments{
\item{LR.ss}{}

\item{LR.ds}{}

\item{prior}{}
}
\value{

}
\description{
Functions to conduct emprical cross entropy calculations these functions are
a reworking of Danial Ramos' ECE functions for Matlab (C) David Lucy 22 June
2010 uses the \code{gpava} function from the \code{isotone} package as a calibrated set
is needed calculates ECE according to Ramos (various) but mostly Ramos IEEE
2008 Equation 15 has a default prior of 99 bins
}
\author{
David Lucy
}
