% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLR.R
\name{calcLR}
\alias{calcLR}
\alias{calcLR_MVN}
\title{Calculate the likelihood ratio}
\usage{
calcLR(control, recovered, background, method = c("mvn", "kde", "lindley"))

calcLR_MVN(control, recovered, background)
}
\arguments{
\item{control}{a \code{compitem} object with the control item information.}

\item{recovered}{a \code{compitem} object with the recovered item information.}

\item{background}{a \code{compcovar} object with the population information.}

\item{method}{a choice of the method used to calculate the LR. Presently there
are three methods, \code{"mvn"} - multivariate normal approximation, \code{"kde"} - (multivariate) kernel
density estimates and \code{"lindely"} which uses the method published by Lindley (1977).}
}
\value{
an estimate of the likelihood ratio
}
\description{
Takes a \code{compitem} object which represents some control item, and a
\code{compitem} object which represents a recovered item, then uses information
from a \code{compcovar} object, which represents the information from the
population, to calculate a likelihood ratio (LR) as a measure of the evidence
given by the observations for the same/different source propositions.
}
\section{Functions}{
\itemize{
\item \code{calcLR_MVN}: Use a multivariate normal approximation to compute the LR
}}

\references{
Aitken, C.G.G. & Lucy, D. (2004) Evaluation of trace evidence in the form of multivariate data. \emph{Applied Statistics}: \bold{53}(1); 109-122.
}
