% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camp3D.R
\name{camp3D}
\alias{camp3D}
\title{camp3D}
\usage{
camp3D(genticmap, len, num)
}
\arguments{
\item{len}{Number of gene fragment groups}

\item{num}{Gene fragment number}
}
\value{
Colinear comparison of maps
}
\description{
camp3D
}
