% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison-package.R
\docType{package}
\name{comparison-package}
\alias{comparison}
\alias{comparison-package}
\title{comparison: Multivariate Likelihood Ratio Calculation and Evaluation}
\description{
Functions for calculating and evaluating likelihood ratios from uni/multivariate continuous observations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jmcurran/comparison}
  \item Report bugs at \url{https://github.com/jmcurran/comparison/issues}
}

}
\author{
\strong{Maintainer}: James Curran \email{j.curran@auckland.ac.nz}

Authors:
\itemize{
  \item David Lucy \email{d.lucy@lancaster.ac.uk}
  \item Agnieszka Martyna \email{ag.rzepecka@gmail.com}
}

}
\keyword{internal}
