% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison.R
\name{load_file}
\alias{load_file}
\title{Generates ggplots by comparing among different user-defined groups.}
\usage{
load_file(file, group_by = NULL)
}
\arguments{
\item{file}{Path to the input file}

\item{group_by}{user-defined groups. By default, it will be gender.}
}
\value{
\code{load_file()} returns 2 ggplots objects \code{transit} and \code{demo}
}
\description{
\code{load_file()} loads .csv file that contains \code{u_egdata.base_job_tran_1step} column fields.
The function returns 2 ggplots:
\enumerate{
\item \code{transit} - Box plot for the time it takes to transit to various industries by user-defined groups
\item \verb{$demo} - Bar plot for the distribution of user-defined groups across various industries
}
}
\examples{

#load library
library('comparison')

#calling load_file function
test <- load_file("./dummy_data.csv", group_by = 'age_bracket')

#Plot Box plot for the time it takes to transit to various industries by different age-bracket groups
test$transit

#Bar plot for the distribution of different age-bracket groups across various industries
test$demo
}
\author{
Kai Wei Tan \href{mailto:kaitan@linkedin.com}{kaitan@linkedin.com}
}
