% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LR_comparison_KDE.r
\name{LR.KDE.function}
\alias{LR.KDE.function}
\title{i' runs through all objects from the population}
\usage{
LR.KDE.function(
  y.mean.1,
  y.mean.2,
  y.star,
  U,
  C,
  h,
  population,
  variables,
  p,
  m,
  n.1,
  n.2
)
}
\description{
i' runs through all objects from the population
}
