% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCompItem.R
\name{makeCompItem}
\alias{makeCompItem}
\title{Create a \code{compitem} object.}
\usage{
makeCompItem(x, ...)
}
\arguments{
\item{x}{a \code{matrix} or \code{data.frame} or a \code{formula}}

\item{\dots}{other arguments that may be passed to the function,
primarily a \code{data.frame} when the formula interface is used}
}
\value{
an object of class \code{compitem}
}
\description{
This function creates a \code{compitem} from a set of observations
on items to be deemed control, or a recovered, items. For example,
a set of elemental concentration measurements on a sample of glass
fragments taken from a crime scene source such as a window.
}
\examples{
# load Greg Zadora's glass data
data(glass)

# calculate a compitem object representing the control item
controlMeasurements = glass[glass$item == "s1", c("logKO", "logCaO", "logFeO")]
control = makeCompItem(controlMeasurements)
}
