% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.R
\name{outlier}
\alias{outlier}
\title{outlier}
\usage{
outlier(genticmap, len, num)
}
\arguments{
\item{len}{Number of gene fragment groups}

\item{num}{Gene fragment number}
}
\value{
colinear comparison of genetic and physical maps ,T test and Error mark
}
\description{
outlier
}
