% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ece.R
\name{plot.ece}
\alias{plot.ece}
\title{An S3 plot method for objects of class \code{ece}}
\usage{
\method{plot}{ece}(x, ...)
}
\arguments{
\item{x}{an S3 object of class \code{ece} which is generated from \code{\link[=calc.ece]{calc.ece()}}.}

\item{\dots}{other arguments that are passed to the \code{plot} generic.}
}
\description{
An S3 plot method for objects of class \code{ece}
}
\seealso{
\code{\link[=calc.ece]{calc.ece()}}
}
\author{
David Lucy
}
