%\name{plot-methods}
\name{plot-method}
\docType{methods}
%\alias{plot-methods}
\alias{plot-method}
\alias{plot,ANY-method}
\alias{plot,ece-method}
\title{Methods for \code{plot} in Package \pkg{graphics}}
\description{Methods for function \code{plot} in package \pkg{graphics}}

\section{Methods}{
\describe{
\item{\code{signature(x = "ANY")}}{
%%  ~~describe this method here~~
}
\item{\code{signature(x = "ece")}}{
%%  ~~describe this method here~~
}
}}

\section{Useage}{plot(x)	# where x is an ece object generated from \code{\link{calc.ece}}}

\references{Ramos, D. & Gonzalez-Rodriguez, J. (2008) Cross-entropy analysis of the information in forensic speaker recognition; IEEE Odyssey.

Zadora, G. & Ramos, D. (2010) Evaluation of glass samples for forensic purposes - an application of likelihood ratio model and information-theoretical approach. \emph{Chemometrics and Intelligent Laboratory}: \bold{102}; 63-83. 
}

\seealso{
\code{\link[isotone]{gpava}}\cr
\code{\linkS4class{ece}}\cr
\code{\link{calibrate.set}}\cr
\code{\link{calc.ece}}
}

\keyword{methods}
%\keyword{ ~~ other possible keyword(s) ~~ }



