% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.compitem.R
\name{print.compitem}
\alias{print.compitem}
\title{S3 method for class \code{compitem}}
\usage{
\method{print}{compitem}(x, ...)
}
\arguments{
\item{x}{an object of class \code{compitem} created by \code{\link[=makeCompItem]{makeCompItem()}}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
S3 method for class \code{compitem}
}
