% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter.R
\name{scatter}
\alias{scatter}
\title{scatter}
\usage{
scatter(genticmap, len, num)
}
\arguments{
\item{len}{Number of gene fragment groups}

\item{num}{Gene fragment number}
}
\value{
Scatter plot of genetic map and physical map
}
\description{
scatter
}
