% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter3D.R
\name{scatter3D}
\alias{scatter3D}
\title{scatter3D}
\usage{
scatter3D(genticmap, len, num)
}
\arguments{
\item{len}{Number of gene fragment groups}

\item{num}{Gene fragment number}
}
\value{
Scatter plot of maps
}
\description{
scatter3D
}
