% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_formula.R
\name{scatter_formula}
\alias{scatter_formula}
\title{scatter_formula}
\usage{
scatter_formula(genticmap, len, num)
}
\arguments{
\item{len}{Number of gene fragment groups}

\item{num}{Gene fragment number}
}
\value{
Scatter plot of genetic map and physical map ,Regression formula and p-value
}
\description{
scatter_formula
}
