% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two.level.density.LR.R
\name{two.level.density.LR}
\alias{two.level.density.LR}
\title{Calculate the likelihood ratio using multivariate KDEs}
\usage{
two.level.density.LR(control, recovered, background)
}
\arguments{
\item{control}{}

\item{recovered}{}

\item{background}{a compcovar object calculated from the observavions of the population as a whole - calculated from the}
}
\value{

}
\description{
Calculates the likelihood ratio for a multivariate random effects with
between items modelled as kernel densities This function could still do with
being made a bit quicker I have tried using apply() type formulations where
appropriate but they are slightly slower than the counted iterations - I
suspect the fact that they are dealing with matrix operations has something
to do with it we're using the prod(eigen(A)$values) form rather than det(A)
as it seems to be a little more reliable as some of the matricies tend to
singularity REQUIRES control - a compitem object calculated from the
observations from the item considered to be the control item - calculated
from two.level.comparison.items() from the file items_two_level.r recovered -
a compitem object calculated from the observations from the item considered
to be the recovered item - calculated from two.level.comparison.items() from
the file items_two_level.r
}
