% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two.level.lindley.LR.R
\name{two.level.lindley.LR}
\alias{two.level.lindley.LR}
\title{Calculate the univariate LR}
\usage{
two.level.lindley.LR(control, recovered, background)
}
\arguments{
\item{background}{a compcovar object calculated from the observavions of the
population as a whole - calculated from the two.level.components() function
from the file components_two_level.r RETURNS LR - an estimate of the
likelihood ratio}
}
\value{

}
\description{
Calculates the likelihood ratio for a univariate random effects with between
items modelled normal this is taken from Lindley's 1977 work and really forms
the precursor to the rest of this package REQUIRES control - a compitem
object calculated from the observations from the item considered to be the
control item - calculated from two.level.comparison.items() from the file
items_two_level.r recovered - a compitem object calculated from the
observations from the item considered to be the recovered item - calculated
from two.level.comparison.items() from the file items_two_level.r
}
