% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two.level.normal.LR.R
\name{two.level.normal.LR}
\alias{two.level.normal.LR}
\title{Calculate a two-level likelihood ratio}
\usage{
two.level.normal.LR(control, recovered, background)
}
\arguments{
\item{control}{}

\item{recovered}{}

\item{background}{a \code{compcovar} object calculated from the observations of the population as a whole -
calculated from the two.level.components() function from the file
components_two_level.r RETURNS LR - an estimate of the likelihood ratio ' '}
}
\value{

}
\description{
Calculates the likelihood ratio for a multivariate random effect model with
between items modelled normal rather than kernal REQUIRES control - a
compitem object calculated from the observations from the item considered to
be the control item - calculated from two.level.comparison.items() from the
file items_two_level.r recovered - a compitem object calculated from the
observations from the item considered to be the recovered item - calculated
from two.level.comparison.items() from the file items_two_level.r
}
