% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UC.R
\name{UC}
\alias{UC}
\title{mean.all' corresponds to the vector of means of all variables calculated using all measurements for all objects
object.mean' corresponds to a matrix of means of all variables calculated using all measurements for each object
i' runs through all objects from the population
j' runs through n measurements for the chosen ith object}
\usage{
UC(population, variables, p, n)
}
\description{
mean.all' corresponds to the vector of means of all variables calculated using all measurements for all objects
object.mean' corresponds to a matrix of means of all variables calculated using all measurements for each object
i' runs through all objects from the population
j' runs through n measurements for the chosen ith object
}
