% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.set.R
\name{calibrate.set}
\alias{calibrate.set}
\title{Calculate the calibrated set of idea LRs}
\usage{
calibrate.set(LR.ss, LR.ds, method = "raw")
}
\arguments{
\item{method}{}
}
\value{

}
\description{
Functions to conduct emprical cross entropy calculations these functions are
a reworking of Danial Ramos' ECE functions for Matlab (C) David Lucy 20 July
2010 calculates and returns the calibrated set of `ideal' LRs from the
observed LRs using the penalised adjacent violators algotrthm This is very
much a rewrite of Nico Brummer's optloglr() function for Matlab, only for R
draws heavily on Nico Brummer's work
}
\author{
David Lucy
}
