% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two.level.comparison.items.R
\name{two.level.comparison.items}
\alias{two.level.comparison.items}
\title{Create a comparison object}
\usage{
two.level.comparison.items(dat, data.columns)
}
\arguments{
\item{dat}{}

\item{data.columns}{}
}
\value{

}
\description{
Makes a comparison.items object. A comparison items object has the data from
the item to be compared - typically for any specific comparison problem there
will be two of these objects requires: dat a data frame of variables and
factors data.columns integer vector indicating which columns in dat contain
the measurements returns: item.means real p array where p is the number of
variables item.n integer: number of replicate observations for each item
n.vars integer: number of contnuous variables multivariate logical T/F
indicates whether multivariate or not observed the raw observations in the
form of an i*p matrix where there are i replicated observations and p
variables missing values: traps and removes NAs - issues warnings about NAs
}
