% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two.level.comparison.items.R
\name{two.level.comparison.items}
\alias{two.level.comparison.items}
\title{Create a \code{compitem} object.}
\usage{
two.level.comparison.items(data, data.columns)
}
\arguments{
\item{data}{a \code{matrix} or \code{data.frame} of observed properties from either
the control item, or the recovered item}

\item{data.columns}{vector of integers giving which columns in \code{data} are
the observations of the properties}
}
\value{
an object of class \code{compitem}
}
\description{
This function creates a \code{compitem} object from a \code{data.frame} or \code{matrix} of
observations from an item to be deemed a control, or a recovered, item.
}
\examples{
# load Greg Zadora's glass data
data(glass)

# calculate a compitem object representing the control item
control = two.level.comparison.items(glass[1:6,], c(7,8,9))
}
